datascience/datascience/requirements.txt
numpy
scipy
sklearn
django
django-rest-framework